I think Accounts are now finished.

+++ app/src/controller/AccountController.php
@@ -30,7 +30,8 @@ class AccountController extends Controller
                 echo $view->addData('accounts', $accounts)->render();
             } catch (BankException $ex) {
                 if ($ex->getCode() == 9) {
-                    $this->redirect('accountCreate');
+                    $view = new View('accountCreate');
+                    echo $view->addData('noAccounts',"TRUE")->render();
                 } else {
                     $view = new View('exception');
                     echo $view->addData("exception", $ex)->addData("back", "Home")->render();
@@ -43,27 +44,44 @@ class AccountController extends Controller
 
     /**
      * Account Create action
-     * @throws BankException
      */
     public function createAction()
     {
         session_start();
-        if (isset($_POST['create'])) {
-            $account = new AccountModel();
-            $account->setType($_POST['accountType']);
-            $account->setUser($_POST['userID']);
-            $account->setBalance(0.0);
-            $account->setDateStarted(date("d/m/y"));
-            $account->save();
-            if(!$account)
-            {
-                throw new BankException(0);
+        if (isset($_SESSION['userName'])) {
+            if (isset($_POST['create'])) {
+                try {
+                    if ($_POST['accountType'] == "undefined") {
+                        throw new BankException(99, "You must select an account type");
+                    }
+                    $account = new AccountModel();
+                    $account->setType($_POST['accountType']);
+                    if (isset($_POST['userId'])) {
+                        $account->setUser($_POST['userId']);
+                    } else {
+                        $account->setUser($_SESSION['userId']);
+                    }
+                    $account->setBalance(0.0);
+                    $account->setDateStarted(date("d/m/Y"));
+                    $account->save();
+                    if (!$account) {
+                        throw new BankException(0);
+                    }
+                    if ($_POST['makeDeposit']) {
+                        $view = new View('accountDeposit');
+                        echo $view->addData("id",$_SESSION['userId'])->render();
+                    }
+                    $this->redirect('accountIndex');
+                } catch (BankException $ex) {
+                    $view = new View('exception');
+                    echo $view->addData("exception", $ex)->addData("back", "Home")->render();
+                }
+            } else {
+                $view = new View('accountCreate');
+                echo $view->render();
             }
-            $view = new View('accountCreate');
-            echo $view->render();
-        } else{
-            $view = new View('accountCreate');
-            echo $view->render();
+        } else {
+            $this->redirect('Home');
         }
     }
 
@@ -71,33 +89,38 @@ class AccountController extends Controller
      * Account Delete action
      *
      * @param int $id Account id to be deleted
-     * @throws BankException
      */
     public function deleteAction($id)
     {
         session_start();
-        try {
-            (new AccountModel())->load($id)->delete();
-        } catch (BankException $e) {
-            throw new BankException('Failed to load account');
+        if ($_SESSION['userName'] == "admin") {
+            try {
+                (new AccountModel())->load($id)->delete();
+            } catch (BankException $ex) {
+                $view = new View('exception');
+                echo $view->addData("exception", $ex)->addData("back", "accountIndex")->render();
+            }
+            $view = new View('accountDeleted');
+            echo $view->addData('accountId', $id)->render();
+        } else {
+            $view = new View('exception');
+            echo $view->addData("exception", (new BankException(99,"Please contact us by phone to close an account")))->addData("back", "accountIndex")->render();
         }
-        $view = new View('accountDeleted');
-        echo $view->addData('accountId', $id)->render();
     }
 
     /**
      * Account Update action
      *
      * @param int $id Account id to be updated
-     * @throws BankException
      */
     public function updateAction($id)
     {
         session_start();
         try {
             $account = (new AccountModel())->load($id);
-        } catch (BankException $e) {
-            throw new BankException('Failed to load account');
+        } catch (BankException $ex) {
+            $view = new View('exception');
+            echo $view->addData("exception", $ex)->addData("back", "accountIndex")->render();
         }
         $account->setName('Joe')->save(); // new name will come from Form data
     }
@@ -114,7 +137,7 @@ class AccountController extends Controller
                     throw new BankException(0);
                 }
                 $view = new View('accountDeposited');
-                echo $view->addData("amount", $_POST['depositAmount'])->addData("account", $account)->addData("back", "accountIndex")->render();
+                echo $view->addData("amount", $_POST['depositAmount'])->addData("account", $account)->render();
             } catch (BankException $ex) {
                 $view = new View('exception');
                 echo $view->addData("exception", $ex)->addData("back", "accountIndex")->render();
@@ -130,16 +153,20 @@ class AccountController extends Controller
     {
         session_start();
         if (isset($_POST['withdraw'])) {
-            $account = (new AccountModel())->load($id);
-            $account->withdraw($_POST['withdrawalAmount']);
-            $account->save();
-            if(!$account)
-            {
-                throw new BankException('Failed to load account');
+            try {
+                $account = (new AccountModel())->load($id);
+                $account->withdraw($_POST['withdrawalAmount']);
+                $account->save();
+                if (!$account) {
+                    throw new BankException(0);
+                }
+                $view = new View('accountWithdrawn');
+                echo $view->addData("amount", $_POST['withdrawalAmount'])->addData("account", $account)->render();
+            } catch (BankException $ex) {
+                $view = new View('exception');
+                echo $view->addData("exception", $ex)->addData("back", "accountIndex")->render();
             }
-            $view = new View('accountWithdraw');
-            echo $view->render();
-        } else{
+        } else {
             $view = new View('accountWithdraw');
             echo $view->render();
         }
